better portability of utils and parser
