+++ hangupsbot/parsers/__init__.py
@@ -1,0 +1,1 @@
+from parsers.kludgy_html_parser import simple_parse_to_segments
